% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateQuantilesWide.R
\name{CalculateQuantilesWide}
\alias{CalculateQuantilesWide}
\title{Calculate quantile cutoffs in a wide format}
\usage{
CalculateQuantilesWide(nums.c, ntiles = 4, verbose = FALSE, ...)
}
\arguments{
\item{nums.c}{Numeric vector.}

\item{ntiles}{Integer.}

\item{verbose}{Boolean.}

\item{...}{Other arguments passed to stats:quantile}
}
\value{
A tibble containing the quantile cutoffs in a single row.
}
\description{
Calculate quantile cutoffs in a wide format
}
\examples{
CalculateQuantilesWide(seq(1, 100))
}
